adding feature to inspect userOp.paymasterAndData
To create more explicit UX, and force the user to opt-in that session
keys can use paymaster funds, I added an additional config sstore
mapping as well as the logic, that smartsession inspects
userop.paymasterAndData. if this field is not zero, at least one userOp
must run. this may be a yesPolicy to get the same behavior as before
this PR, but it can be a paymastter policy that inspects the
paymasterAndData for conditions